[REFACTOR] 연산로직에서 switch/case 구문제거
