+++ src/main/java/calculator/Calculator.java
@@ -36,51 +36,4 @@ public class Calculator {
 				.orElse(new Term(0))
 				.getValue();
 	}
-
-
-	private class Term {
-		private int value;
-
-		private Operator operator;
-
-		public Term(int value) {
-
-			this(Operator.ADD, value);
-		}
-
-		public Term(Operator operator, int value) {
-			this.operator = operator;
-			this.value = value;
-		}
-
-		public int getValue(){
-			return value;
-		}
-
-		public Operator getOperator(){
-			return operator;
-		}
-
-		public Term reduce(Term otherTerm){
-			int initial = this.value;
-			switch(otherTerm.getOperator()){
-				case ADD:
-					initial += otherTerm.getValue();
-					break;
-				case DIFFERENCE:
-					initial -= otherTerm.getValue();
-					break;
-				case MULTIPLY:
-					initial *= otherTerm.getValue();
-					break;
-				case DIVISION:
-					initial /= otherTerm.getValue();
-					break;
-				default:
-					throw new IllegalStateException("방법을 알 수 없는 연산자");
-			}
-
-			return new Term(initial);
-		}
-	}
 }
