+++ src/main/java/calculator/OperationFunc.java
@@ -1,0 +1,6 @@
+package calculator;
+
+@FunctionalInterface
+public interface OperationFunc {
+	int execute(int a, int b);
+}
